Moved backend logic to new location
